[python] extract annotation_utils module from python_annotation.h
Move six pure, state-free helpers and the `InferResult` enum out of
python_annotation.h into a new python_annotation/annotation_utils
module:

  - InferResult (enum)
  - has_annotation
  - has_return_none
  - get_type_from_json
  - get_base_var_name
  - invert_substrings
  - infer_type_from_default_arg_shape

Bodies are transcribed verbatim and now take `const nlohmann::json &`
directly (`python_annotation` was only ever instantiated with
`nlohmann::json`). A block of `using` declarations preserves the
unqualified names inside python_annotation.h so all class-internal
call sites remain byte-identical. The single external caller —
`converter_stmt.cpp` — now reaches the helper through
`python_annotation_utils::` instead of the static class method.
No semantic change to inference, no API change for the class.
