+++ src/python-frontend/CMakeLists.txt
@@ -39,6 +39,7 @@ add_library(pythonfrontend STATIC
             module_manager.cpp
             module_locator.cpp
             python_annotation/annotation_intrinsics.cpp
+            python_annotation/annotation_utils.cpp
             symbol_id.cpp
             type_handler.cpp
             cmath_lowering_policy.cpp
